+++ frontend/src/components/Section.tsx
@@ -1,0 +1,53 @@
+import { Box, Flex, Heading, Separator } from '@radix-ui/themes';
+import type { ReactNode } from 'react';
+
+interface SectionCardProps {
+  // Plain string renders as the bold heading; pass a node for a custom header row.
+  title: ReactNode;
+  // Optional icon shown to the left of a string title.
+  icon?: ReactNode;
+  // Optional right-aligned header content (e.g. a "Show all" button or count badge).
+  action?: ReactNode;
+  children: ReactNode;
+  // Card height — e.g. "100%" to equalise cards sharing a Grid row.
+  h?: string | number;
+  // When true, the body scrolls internally (up to SCROLL_MAX_HEIGHT) while the header stays
+  // pinned above it — used by the long, frequently-polled run-status tables.
+  scroll?: boolean;
+}
+
+const SCROLL_MAX_HEIGHT = 600;
+
+// A bordered card with a titled header strip and a full-width separator. Replaces the repeated
+// "panel with a heading bar" pattern (was Mantine's Card.Section).
+export function Section({ title, icon, action, children, h, scroll }: SectionCardProps) {
+  return (
+    <Box
+      style={{
+        height: h,
+        overflow: 'hidden',
+      }}
+    >
+      <Flex justify="between" align="center">
+        {typeof title === 'string' ? (
+          <Flex align="center" gap="2" pb="3" style={{borderBottom:"2px solid var(--green-9)"}} >
+            {icon}
+            <Heading as="h3" size="3" >
+              {title}
+            </Heading>
+          </Flex>
+        ) : (
+          title
+        )}
+        {action}
+      </Flex>
+      <Box
+        px="6"
+        py="6"
+        style={scroll ? { maxHeight: SCROLL_MAX_HEIGHT, overflowY: 'auto' } : undefined}
+      >
+        {children}
+      </Box>
+    </Box>
+  );
+}
